Make MPI wrapper usable in serial
